+++ relay-server/src/services/projects/cache/project.rs
@@ -9,7 +9,6 @@ use crate::managed::ManagedEnvelope;
 use crate::services::outcome::{DiscardReason, Outcome};
 use crate::services::projects::cache::state::SharedProject;
 use crate::services::projects::project::ProjectState;
-use crate::statsd::RelayTimers;
 use crate::utils::{CheckLimits, Enforcement, EnvelopeLimiter};
 
 /// A loaded project.
@@ -88,9 +87,7 @@ impl<'a> Project<'a> {
 
         // If we can extract spans from the event, we want to try and count the number of nested
         // spans to correctly emit negative outcomes in case the transaction itself is dropped.
-        relay_statsd::metric!(timer(RelayTimers::CheckNestedSpans), {
-            sync_spans_to_enforcement(&envelope, &mut enforcement);
-        });
+        sync_spans_to_enforcement(&mut envelope, &mut enforcement);
 
         enforcement.apply_with_outcomes(&mut envelope);
 
@@ -135,12 +132,17 @@ pub struct CheckedEnvelope {
 /// On the fast path of rate limiting, we do not have nested spans of a transaction extracted
 /// as top-level spans, thus if we limited a transaction, we want to count and emit negative
 /// outcomes for each of the spans nested inside that transaction.
-fn sync_spans_to_enforcement(envelope: &ManagedEnvelope, enforcement: &mut Enforcement) {
+fn sync_spans_to_enforcement(envelope: &mut ManagedEnvelope, enforcement: &mut Enforcement) {
     if !enforcement.is_event_active() {
         return;
     }
 
-    let spans_count = count_nested_spans(envelope);
+    let spans_count = envelope
+        .envelope_mut()
+        .items_mut()
+        .find(|item| *item.ty() == ItemType::Transaction && !item.spans_extracted())
+        .map_or(0, |item| 1 + item.ensure_span_count());
+
     if spans_count == 0 {
         return;
     }
@@ -149,28 +151,12 @@ fn sync_spans_to_enforcement(envelope: &ManagedEnvelope, enforcement: &mut Enfor
         enforcement.spans = enforcement.event.clone_for(DataCategory::Span, spans_count);
     }
 
+    // TODO(follow-up): Do not manually enforce, rely on quantities() instead.
     if enforcement.event_indexed.is_active() {
         enforcement.spans_indexed = enforcement
             .event_indexed
             .clone_for(DataCategory::SpanIndexed, spans_count);
     }
-}
-
-/// Counts the nested spans inside the first transaction envelope item inside the [`Envelope`](crate::envelope::Envelope).
-fn count_nested_spans(envelope: &ManagedEnvelope) -> usize {
-    #[derive(Debug, serde::Deserialize)]
-    struct PartialEvent {
-        spans: crate::utils::SeqCount,
-    }
-
-    envelope
-        .envelope()
-        .items()
-        .find(|item| *item.ty() == ItemType::Transaction && !item.spans_extracted())
-        .and_then(|item| serde_json::from_slice::<PartialEvent>(&item.payload()).ok())
-        // We do + 1, since we count the transaction itself because it will be extracted
-        // as a span and counted during the slow path of rate limiting.
-        .map_or(0, |event| event.spans.0 + 1)
 }
 
 #[cfg(test)]
@@ -211,6 +197,15 @@ mod tests {
             .unwrap();
 
         RequestMeta::new(dsn)
+    }
+
+    fn get_span_count(managed_envelope: &ManagedEnvelope) -> usize {
+        managed_envelope
+            .envelope()
+            .items()
+            .next()
+            .unwrap()
+            .span_count()
     }
 
     #[tokio::test]
@@ -269,7 +264,9 @@ mod tests {
 
         let managed_envelope = ManagedEnvelope::new(envelope, outcome_aggregator.clone());
 
+        assert_eq!(get_span_count(&managed_envelope), 0); // not written yet
         project.check_envelope(managed_envelope).await.unwrap();
+
         drop(outcome_aggregator);
 
         let expected = [
@@ -277,6 +274,69 @@ mod tests {
             (DataCategory::TransactionIndexed, 1),
             (DataCategory::Span, 3),
             (DataCategory::SpanIndexed, 3),
+        ];
+
+        for (expected_category, expected_quantity) in expected {
+            let outcome = outcome_aggregator_rx.recv().await.unwrap();
+            assert_eq!(outcome.category, expected_category);
+            assert_eq!(outcome.quantity, expected_quantity);
+        }
+    }
+
+    #[tokio::test]
+    async fn test_track_nested_spans_outcomes_predefined() {
+        let config = Default::default();
+        let project = create_project(
+            &config,
+            Some(json!({
+                "quotas": [{
+                   "id": "foo",
+                   "categories": ["transaction"],
+                   "window": 3600,
+                   "limit": 0,
+                   "reasonCode": "foo",
+               }]
+            })),
+        );
+
+        let mut envelope = Envelope::from_request(Some(EventId::new()), request_meta());
+
+        let mut transaction = Item::new(ItemType::Transaction);
+        transaction.set_span_count(Some(666));
+        transaction.set_payload(
+            ContentType::Json,
+            r#"{
+  "event_id": "52df9022835246eeb317dbd739ccd059",
+  "type": "transaction",
+  "transaction": "I have a stale timestamp, but I'm recent!",
+  "start_timestamp": 1,
+  "timestamp": 2,
+  "contexts": {
+    "trace": {
+      "trace_id": "ff62a8b040f340bda5d830223def1d81",
+      "span_id": "bd429c44b67a3eb4"
+    }
+  },
+  "spans": []
+}"#,
+        );
+
+        envelope.add_item(transaction);
+
+        let (outcome_aggregator, mut outcome_aggregator_rx) = relay_system::Addr::custom();
+
+        let managed_envelope = ManagedEnvelope::new(envelope, outcome_aggregator.clone());
+
+        assert_eq!(get_span_count(&managed_envelope), 666);
+        project.check_envelope(managed_envelope).await.unwrap();
+
+        drop(outcome_aggregator);
+
+        let expected = [
+            (DataCategory::Transaction, 1),
+            (DataCategory::TransactionIndexed, 1),
+            (DataCategory::Span, 667),
+            (DataCategory::SpanIndexed, 667),
         ];
 
         for (expected_category, expected_quantity) in expected {